# backend/docs/openapi-spec.yaml
openapi: 3.0.0
info:
  title: Pet Adoption API
  version: 1.0.0
  description: |
    API RESTful para gestión de mascotas disponibles para adopción, pacientes y facturación de visitas.
    Esta API permite buscar mascotas por diferentes criterios, registrar nuevas mascotas,
    gestionar la información de pacientes, y generar facturas detalladas por visita.
  contact:
    name: DUOC UC - CDY2203
    url: https://www.duoc.cl
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Development Server
  - url: https://api.example.com
    description: Production Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token for authentication
  schemas:
    Pet:
      type: object
      required:
        - name
        - species
        - breed
        - age
        - gender
      properties:
        id:
          type: integer
          description: Identificador único de la mascota
          example: 1
        name:
          type: string
          description: Nombre de la mascota
          example: Max
        species:
          type: string
          description: Especie de la mascota
          example: Perro
        breed:
          type: string
          description: Raza de la mascota
          example: Golden Retriever
        age:
          type: integer
          description: Edad de la mascota en años
          example: 3
        gender:
          type: string
          description: Sexo de la mascota
          enum:
            - Macho
            - Hembra
          example: Macho
        location:
          type: string
          description: Ubicación geográfica de la mascota
          example: Santiago
        photos:
          type: array
          description: Lista de URLs de fotos de la mascota
          items:
            type: string
            format: uri
          example:
            - "https://example.com/foto1.jpg"
            - "https://example.com/foto2.jpg"
        status:
          type: string
          description: Estado de disponibilidad
          enum:
            - available
            - adopted
          default: available
          example: available
...
    UserRegistrationRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          description: Nombre de usuario único
          example: recepcionista01
        email:
          type: string
          format: email
          description: Correo electrónico único del usuario
          example: recepcion@duocvet.cl
        password:
          type: string
          format: password
          description: Contraseña en texto plano enviada sólo al registrar; se almacena como hash BCrypt
          example: MiClaveSegura2026!
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          description: Identificador del usuario creado
          example: 7
        username:
          type: string
          description: Nombre de usuario registrado
          example: recepcionista01
        email:
          type: string
          format: email
          description: Correo electrónico registrado
          example: recepcion@duocvet.cl
    Error:
      type: object
      properties:
        message:
          type: string
          description: Mensaje de error
          example: "Error al registrar mascota: Campo requerido"
...
paths:
...
  /users/register:
    post:
      tags:
        - Autenticación
      summary: Registrar nuevo usuario
      description: Crea un usuario nuevo en la base de datos y almacena la contraseña como hash BCrypt. Esta ruta requiere autenticación JWT válida.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
      responses:
        '201':
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Datos inválidos o usuario duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autenticado
        '403':
          description: Token inválido o expirado
...
  /pets:
    get:
      tags:
        - Mascotas
      summary: Obtener todas las mascotas
      description: Recupera la lista completa de mascotas registradas en el sistema
      responses:
        '200':
          description: Lista de mascotas obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
        '500':
          description: Error interno del servidor

    post:
      tags:
        - Mascotas
      summary: Registrar nueva mascota
      description: Crea un nuevo registro de mascota disponible para adopción
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
      responses:
        '201':
          description: Mascota registrada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /pets/{id}:
    get:
      tags:
        - Mascotas
      summary: Obtener mascota por ID
      description: Recupera la información detallada de una mascota específica
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID de la mascota
          example: 1
      responses:
        '200':
          description: Mascota encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '404':
          description: Mascota no encontrada
        '500':
          description: Error interno del servidor

    put:
      tags:
        - Mascotas
      summary: Actualizar mascota
      description: Actualiza la información de una mascota existente (cambiar estado a adoptada, etc.)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID de la mascota a actualizar
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
      responses:
        '200':
          description: Mascota actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Mascota no encontrada

    delete:
      tags:
        - Mascotas
      summary: Eliminar mascota
      description: Elimina un registro de mascota del sistema
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID de la mascota a eliminar
          example: 1
      responses:
        '200':
          description: Mascota eliminada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Mascota eliminada exitosamente"
        '404':
          description: Mascota no encontrada
        '400':
          description: Error al eliminar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /pets/available:
    get:
      tags:
        - Mascotas
      summary: Obtener mascotas disponibles
      description: Recupera la lista de mascotas disponibles para adopción
      responses:
        '200':
          description: Lista de mascotas disponibles obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
        '500':
          description: Error interno del servidor
  /pets/search:
    get:
      tags:
        - Mascotas
      summary: Buscar mascotas por criterios
      description: |
        Busca mascotas por uno o varios criterios (especie, género, ubicación, edad).
        Todos los parámetros son opcionales.
      parameters:
        - name: species
          in: query
          required: false
          schema:
            type: string
          description: Filtrar por especie
          example: Perro

        - name: gender
          in: query
          required: false
          schema:
            type: string
            enum:
              - Macho
              - Hembra
          description: Filtrar por sexo
          example: Macho

        - name: location
          in: query
          required: false
          schema:
            type: string
          description: Filtrar por ubicación
          example: Santiago

        - name: age
          in: query
          required: false
          schema:
            type: integer
          description: Filtrar por edad exacta
          example: 3

        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - available
              - adopted
            default: available
          description: Filtrar por estado de disponibilidad
          example: available

      responses:
        '200':
          description: Búsqueda completada exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
        '500':
          description: Error interno del servidor
...